requirements.txt
tabulate
pillow
pypdf2